Add an option on Uno boards to enable a 4 MHz clock signal on pin 3
This clock signal can be used to reprogram the fuses of an AVR that
has been set wrong (ie. when an external clock has been configured but
the board/chip does not have one).

This is disabled by default. We check that we compile for known chips
before enabling the clock generation. Maybe there is other chips that
support this code.