Update issue template to reflect change in GitHub's new issue page
At the time the original issue template was created, GitHub always displayed a reminder to read CONTRIBUTING.md with a link above the new issue creation field. Unfortunately, this has since been changed so that a reminder is now shown on the right side of the field only for new contributors or if CONTRIBUTING.md has changed since the user's last contribution, and a less prominent link to the right otherwise.