Hide splash screen asap if not using gui
Fixes #1970

On macOS, when running,

    Arduino.app/Contents/MacOS/Arduino --upload sketch.ino

a splash screen is visible the entire time the upload is running.

This PR hides that splash screen as soon as it’s known for sure that the
GUI should not be displayed. This makes for a slightly better user
experience: there’s still a splash screen, but it only appears briefly at
launch time.

The [SplashScreen docs][] explain:

> The splash screen can be displayed at application startup, before the
> Java Virtual Machine (JVM) starts. ... is closed automatically as soon as
> the first window is displayed by Swing/AWT (may be also closed manually
> using the Java API, see below)

[SplashScreen docs]: https://docs.oracle.com/javase/7/docs/api/java/awt/SplashScreen.html