Run tests with an X virtual framebuffer (1024x768)
The GitHub-hosted runner is a virtual machine hosted by GitHub.

Xvfb on that setup defaults to a virtual screen size of 640x480
(XVFBARGS='-screen 0 640x480x16') which in turn makes the Arduino user
interface tests fail with exceptions like this:

    [junit] org.fest.swing.exception.ActionFailedException: The component to click is out of the boundaries of the screen
    [junit] 	at org.fest.swing.exception.ActionFailedException.actionFailure(ActionFailedException.java:33)
    [junit] 	at org.fest.swing.core.RobotEventGenerator.pressMouse(RobotEventGenerator.java:74)

Setting the xvfb screen size to 1024x768 solves this error.